requirements.txt
flask-restful
python-dotenv
psycopg2-binary
Flask-JWT-Extended
flask-cors
bcrypt